Extrae CountryReinforcement y Regrouping a archivos aparte
