+++ core/src/main/gamelogic/CountryReinforcement.kt
@@ -1,0 +1,14 @@
+package gamelogic
+
+import Country
+import Player
+import PositiveInt
+import gamelogic.occupations.CountryOccupations
+
+class CountryReinforcement(val country: Country, val armies: PositiveInt) {
+    fun apply(player: Player, occupations: CountryOccupations) {
+        CountryIsNotOccupiedByPlayerException(country, player)
+            .assertPlayerOccupiesCountryIn(occupations)
+        occupations.addArmies(country, armies)
+    }
+}
